misc/HtmlTestSuite/results/test-0182.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Lists: test list with textarea as its first child</title>
<link href="https://www.w3.org/TR/CSS22/generate.html#lists" rel="help"/>
<!-- https://bugs.chromium.org/p/chromium/issues/detail?id=767408 -->

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

<style>

ul, textarea {
  font-size: 16px;
}
</style>

</head><body><div id="log"/>

<ul>
  <li id="target1">
    <textarea cols="20" id="target2" rows="3" style="display:block; height:45px">
      hello
    </textarea>
  </li>
</ul>

<script>
test(function() {
  var height1 = document.getElementById("target1").offsetHeight;
  var height2 = document.getElementById("target2").offsetHeight;
  assert_equals(height1, height2, "List marker and textarea should be in the same line, no line-break between them.")
}, "list and block textarea");

</script>
</body></html>
